refactor: checkBasicPermission as util
